+++ src/main/java/com/example/jobis/domain/recruit/controller/dto/response/RecruitListResponse.java
@@ -1,6 +1,7 @@
 package com.example.jobis.domain.recruit.controller.dto.response;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 
 import java.util.List;
@@ -10,4 +11,16 @@ import java.util.List;
 public class RecruitListResponse {
 
     private final List<RecruitResponse> recruitList;
+
+    @Getter
+    @Builder
+    public class RecruitResponse {
+
+        private final Long recruitId;
+        private final String companyName;
+        private final String companyProfileUrl;
+        private final Integer trainPay;
+        private final boolean military;
+        private final List<Long> codeList;
+    }
 }
